Worker refactor: Separate code into multiple files
